Added Mod Creation to HTTP API
