+++ Penumbra/API/Models/NewMod.cs
@@ -1,0 +1,7 @@
+namespace Penumbra.Api.Models
+{
+    public class NewMod
+    {
+        public string name { get; set; }
+    }
+}
